// src/store/action-types/types.ts
export enum Actiontype {
	SET_ALERT = 'SET_ALERT',
	REMOVE_ALERT = 'REMOVE_ALERT',
	OTPSENT_SUCCESS = 'OTPSENT_SUCCESS',
	OTPSENT_FAIL = 'OTPSENT_FAIL',
	LOGIN_SUCCESS = 'LOGIN_SUCCESS',
	LOGIN_FAIL = 'LOGIN_FAIL',
	LOGOUT = 'LOGOUT',
	USER_LOADED = 'USER_LOADED',
	AUTH_ERROR = 'AUTH_ERROR',
	GET_PRODUCT = 'GET_PRODUCT',
	GET_PRODUCTS = 'GET_PRODUCTS',
	GET_ORDERS = 'GET_ORDERS',
	UPDATE_QUANTITY = 'UPDATE_QUANTITY',
	ADD_TO_CART = 'ADD_TO_CART',
	DELETE_CART_ITEM = 'DELETE_CART_ITEM',
	SET_DELIVERY_ADDRESS = 'SET_DELIVERY_ADDRESS',
	UPDATE_FILTERED_PRODUCTS = 'UPDATE_FILTERED_PRODUCTS',
	UPDATE_FILTERED_ARRAY = 'UPDATE_FILTERED_ARRAY',
	FILTER_BY_TYPE = 'FILTER_BY_TYPE',
	PLACE_ORDER = 'PLACE_ORDER',
	SORT_ACTION = 'SORT_ACTION',
	ADD_ADDRESS = 'ADD_ADDRESS',
	CHANGE_LOGIN_MODAL = 'CHANGE_LOGIN_MODAL',
}